fix(nvim/typst): don't let typst-lsp clash with `typst watch`
typst-lsp actually runs typst under the hood, but the config it is run
with cannot be influenced. As such, it always ends up with the default
namign scheme. Previously, `typst watch` as it was called here also used
the default naming scheme. This often resulted in the two overriding
each other.

Normally, this is not an issue, since the two were run with very similar
changes. However, typst-lsp's default arguments also mean that it
doesn't take any `--input` arguments. In this case, when `typst watch`
was run *with* `--input` arguments which affect output, it was very
likely that the PDF viewer flickered back and forth between the 2
versions, which was unpleasing to view and very confusing.

This is fixed by having `typst watch` go out of typst-lsp's way by
having it write to $(pwd)/view.pdf instead, which is also opened in
evince. As such, the two write to different files, and no race condition
is possible anymore, as the file written to by typst-lsp is never
opened.